+++ Modules/Intent.Modules.Common/release-notes.md
@@ -1,3 +1,7 @@
+### Version 3.9.1
+
+-Improvement: Created `ExecutionLifeCycle` static for statically registering actions to be executed during Software Factory runs. Currently used to clear static cache between "hot SF" runs.
+
 ### Version 3.9.0
 
 - Feature: Added support for filtering template instances for registrations.